Wrap main section in content div
I wrapped my main card content in a div with a class in order to
apply universal padding and margin styles to every element nested
within.

Ideally I will refactor the code to include more CSS grid properties,
but with the time given to complete this project I'd like to first
deliver a MVP and then focus on refactoring.