Space bottom buttons
I had to add an ID to the second button in my footer to space it
appropriately.  If time permits, I want to go back and refactor this
code to use only CSS grid properties.  That involves including a
sub-grid for the main card content.

I want to spend the remaining time working on adding a tooltip tail to
the top box.